[squeeze] Use previsouly built iptables packages
We need custom iptables because of an IMQ patch. Unfortuantely, the
version we patch upon (1.4.8) cannot be built with the
linux-libc-dev in
our archive, namely the version from squeeze-backports.

This squeeze-backports version is in because we build pixman and
cairo
from source, to serve as iceweasel build-dependencies.

In the end, there are two solutions (since linux-libc-dev can't be
downgraded) :

  1. port the IMQ patch to a more recent version of iptables, that'd
     support this recent linux-libc-dev

  2. dump in svn the binary iptables packages we've been using so
  far,
     and stop building iptables from source

It's too late to go with the better, #1 solution right now, so #2 it is.



git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/upstream/pkgs@36080 8ebf8b4e-06f3-0310-961f-89f74618be54